# swagger.yaml
openapi: 3.0.3
info:
  title: SPV APIs
  description: |-
    This is a the bloom-request OpenAPI 3.0 specification
  contact:
    email: joe@netti.dev
  version: 0
externalDocs:
  description: Github repo
  url: https://github.com/nettijoe96/spv-bloom
tags:
  - name: Message APIs
    description: APIs to publish and retrieve messages
paths:
  /bloom-request:
    post:
      parameters:
      - in: query
        name: timestamp
        schema:
          type: string
        required: false
      tags:
        - Message APIs
      summary: Retrieve all messages that match bloom filter
      description: Retrieve all messages that match bloom filter
      operationId: bloom-request
      requestBody:
        description: bloom filter hex string
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Bloom"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Messages"
        "400":
          description: Invalid bloom filter hex
        "408":
          description: Request Timeout
  /publish:
    post:
        parameters:
          - in: query
            name: timestamp
            schema:
              type: string
            required: false
        tags:
          - Message APIs
        summary: Publish new messages
        description: Publish new messages
        operationId: publish
        requestBody:
          description: messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Messages"
          required: true
        responses:
          "204":
            description: Successful operation
          "400":
            description: Invalid request body
          "408":
            description: Request Timeout
components:
  schemas:
    Messages:
      type: object
      properties:
        messages:
          type: array
          items:
            type: array
            items:
              type: string
    Bloom:
      type: object
      properties:
        bloom:
          type: string